Extracted views context to separate file
